speedreader: Fix local cargo build
Remove obsolete benchmark target and fix relative path to the
kuchiki dependency so `cargo check` passes in the local crate
directory for the rust code. This facilitates local testing
and comparison between cargo and the gn builds.

For the same reason, add a similar diversion to the ffi crate.
Each top-level Cargo.toml needs to declare depencency patches
separately.