Remove `BraveVpnService` and `BraveVpnWireguardService` from services
This code runs when mini-installer executes for system level installs.
The install worker is executed for new users but also runs on each upgrade.

This is the first part of the fix, helping solve https://github.com/brave/brave-browser/issues/33726

The second part of the fix will be to install the services when needed (and is a bit tricky).