Updated interface IDs for all items in elevation_service_idl.idl
This also created a new interface for Development (not in upstream)

This service is not used by Brave currently, but it's in use by Chrome.
We don't want to have conflicts with the IDs.

Updated IID for elevation service in registration (in install_static)

Updates service naming logic to take channel into account