Add JsSkusPlatformBrowserTest
Add a test that simulates calling the SKUs SDK from javascript,
triggering an error, and having that error caught by javascript.

In order to support this test I needed to modify IsSafeOrigin to strip
the port from the origin when making the comparison against the allowed
list. The reason is that the port is randomly assigned in the test, and
so it never matches the expected port.